fix: moved to mjs and added gitcommit sha as versioning for sw
