+++ apps/web/package.json
@@ -4,9 +4,9 @@
   "private": true,
   "license": "AGPL-3.0",
   "scripts": {
-    "build:sw": "node script/build-sw.js",
+    "build:sw": "node script/build-sw.mjs",
     "build": "lingui compile; pnpm run build:sw; next build",
-    "dev": "lingui compile; pnpm run build:sw; next dev --port 4783",
+    "dev": "lingui compile && pnpm run build:sw && next dev --port 4783",
     "extract": "lingui compile; lingui extract",
     "lint": "eslint . --ext .ts,.tsx",
     "lint:fix": "eslint . --fix --ext .ts,.tsx",
@@ -95,6 +95,7 @@
     "@types/react-dom": "^18.2.10",
     "@types/uuid": "^9.0.4",
     "autoprefixer": "^10.4.16",
+    "dotenv": "^16.3.1",
     "esbuild": "^0.19.4",
     "postcss": "^8.4.31",
     "tailwindcss": "^3.3.3",
